+++ crates/ewwii/src/app.rs
@@ -155,7 +155,8 @@ pub struct App<B: DisplayBackend> {
     /// The user's css provider.
     pub css_provider: gtk4::CssProvider,
     /// This will be set by the plugins.
-    pub custom_css_providers: Vec<gtk4::CssProvider>,
+    pub custom_css_providers: Vec<Option<gtk4::CssProvider>>,
+    pub gdk_display: Option<gtk4::gdk::Display>,
     pub plugin_buffer: plugin::PluginBuffer,
     pub reloading: bool,